Migrate vis_type_table to kibana/new platform  (#63105)
* Move vis_type_table to Kibana Platform

* Adapt mocha tests
* Adapt CODEOWNERS

* Adapt SCSS
