+++ src/legacy/core_plugins/kibana/public/__tests__/vis_type_table/tabified_data.js
