--- src/legacy/core_plugins/vis_type_table/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "table_vis",
-  "version": "kibana"
-}
