--- src/legacy/core_plugins/vis_type_table/public/agg_table/_index.scss
@@ -1,1 +1,0 @@
-@import 'agg_table';
