--- src/legacy/core_plugins/vis_type_table/public/paginated_table/_index.scss
@@ -1,1 +1,0 @@
-@import './table_cell_filter';
