+++ src/plugins/vis_type_table/kibana.json
@@ -1,0 +1,11 @@
+{
+  "id": "visTypeTable",
+  "version": "kibana",
+  "server": true,
+  "ui": true,
+  "requiredPlugins": [
+    "expressions",
+    "visualizations",
+    "data"
+  ]
+}
