+++ src/plugins/vis_type_table/public/_table_vis.scss
