+++ src/plugins/vis_type_table/public/agg_table/_agg_table.scss
