+++ src/plugins/vis_type_table/public/agg_table/_index.scss
@@ -1,0 +1,1 @@
+@import './agg_table';
