+++ src/plugins/vis_type_table/public/agg_table/agg_table.html
