+++ src/plugins/vis_type_table/public/components/table_vis_options.tsx
@@ -24,12 +24,8 @@ import { i18n } from '@kbn/i18n';
 import { FormattedMessage } from '@kbn/i18n/react';
 
 import { VisOptionsProps } from 'src/plugins/vis_default_editor/public';
-import { search } from '../../../../../plugins/data/public';
-import {
-  SwitchOption,
-  SelectOption,
-  NumberInputOption,
-} from '../../../../../plugins/charts/public';
+import { search } from '../../../data/public';
+import { SwitchOption, SelectOption, NumberInputOption } from '../../../charts/public';
 import { TableVisParams } from '../types';
 import { totalAggregations } from './utils';
