+++ src/plugins/vis_type_table/public/components/utils.ts
