+++ src/plugins/vis_type_table/public/index.scss
@@ -1,5 +1,3 @@
-@import 'src/legacy/ui/public/styles/styling_constants';
-
 // Prefix all styles with "tbv" to avoid conflicts.
 // Examples
 //   tbvChart
@@ -7,6 +5,6 @@
 //   tbvChart__legend--small
 //   tbvChart__legend-isLoading
 
-@import 'agg_table/index';
-@import 'paginated_table/index';
+@import './agg_table/index';
+@import './paginated_table/index';
 @import './table_vis';
