+++ src/plugins/vis_type_table/public/paginated_table/_index.scss
@@ -1,0 +1,1 @@
+@import './_table_cell_filter';
