+++ src/plugins/vis_type_table/public/paginated_table/_table_cell_filter.scss
