+++ src/plugins/vis_type_table/public/paginated_table/paginated_table.html
