+++ src/plugins/vis_type_table/public/paginated_table/table_cell_filter.html
