+++ src/plugins/vis_type_table/public/table_vis.html
