+++ src/plugins/vis_type_table/public/table_vis_fn.test.ts
@@ -21,7 +21,7 @@ import { createTableVisFn } from './table_vis_fn';
 import { tableVisResponseHandler } from './table_vis_response_handler';
 
 // eslint-disable-next-line
-import { functionWrapper } from '../../../../plugins/expressions/common/expression_functions/specs/tests/utils';
+import { functionWrapper } from '../../expressions/common/expression_functions/specs/tests/utils';
 
 jest.mock('./table_vis_response_handler', () => ({
   tableVisResponseHandler: jest.fn().mockReturnValue({
