+++ src/plugins/vis_type_table/public/table_vis_response_handler.ts
